Update DBStatement for SQLite #9
I had to split DBStatement for SQLite into two branches: for PHP 7 and for PHP 8, otherwise it does not allow inheriting the PDOStatement class.
